1.0.4-update

+++ src/main/java/com/vbyte/decisionengine/rules/Operator.java
@@ -8,51 +8,51 @@ public final class Operator {
     /**
      * 大于
      * */
-    public static final byte GREATER_THAN = 0;
+    public static final int GREATER_THAN = 0;
 
     /**
      * 小于
      * */
-    public static final byte LESS_THAN = 1;
+    public static final int LESS_THAN = 1;
 
     /**
      * 等于
      * */
-    public static final byte BE_EQUAL_TO = 2;
+    public static final int BE_EQUAL_TO = 2;
 
     /**
      * 不等于
      * */
-    public static final byte NOT_EQUAL_TO = 3;
+    public static final int NOT_EQUAL_TO = 3;
 
     /**
      * 大于等于
      * */
-    public static final byte GREATER_BE_EQUAL = 4;
+    public static final int GREATER_BE_EQUAL = 4;
 
     /**
      * 小于等于
      * */
-    public static final byte LESS_BE_EQUAL = 5;
+    public static final int LESS_BE_EQUAL = 5;
 
     /**
      * 包含
      * */
-    public static final byte CONTAIN = 6;
+    public static final int CONTAIN = 6;
 
     /**
      * 不包含
      * */
-    public static final byte NOT_CONTAIN = 7;
+    public static final int NOT_CONTAIN = 7;
 
     /**
      * 与
      * */
-    public static final byte AND = 8;
+    public static final int AND = 8;
 
     /**
      * 或
      * */
-    public static final byte OR = 9;
+    public static final int OR = 9;
 
 }
